+++ src/cetech/playground/private/log_view.cpp
@@ -157,7 +157,7 @@ static void ui_log_items() {
     ImGui::EndChild();
 }
 
-static void on_debugui(uint64_t bus_name,
+static void on_debugui(uint32_t bus_name,
                        void *event) {
 
     if (ct_debugui_a0.BeginDock(WINDOW_NAME,
@@ -172,7 +172,7 @@ static void on_debugui(uint64_t bus_name,
 
 }
 
-static void on_menu_window(uint64_t bus_name,
+static void on_menu_window(uint32_t bus_name,
                            void *event) {
     ct_debugui_a0.MenuItem2(WINDOW_NAME, NULL, &_G.visible, true);
 }
